additional warning not previously flagged

+++ robot_calibration/test/chain_manager_tests.cpp
@@ -23,10 +23,10 @@ TEST(ChainManagerTests, test_rosparam_loading)
   robot_calibration::ChainManager manager(nh, 0.001);
 
   std::vector<std::string> chain_names = manager.getChains();
-  EXPECT_EQ(2, chain_names.size());
+  EXPECT_EQ(static_cast<size_t>(2), chain_names.size());
 
   std::vector<std::string> joint_names = manager.getChainJointNames("arm");
-  EXPECT_EQ(7, joint_names.size());
+  EXPECT_EQ(static_cast<size_t>(7), joint_names.size());
   EXPECT_EQ("arm_lift_joint", joint_names[0]);
   EXPECT_EQ("arm_shoulder_pan_joint", joint_names[1]);
   EXPECT_EQ("arm_shoulder_lift_joint", joint_names[2]);
@@ -36,10 +36,10 @@ TEST(ChainManagerTests, test_rosparam_loading)
   EXPECT_EQ("arm_wrist_roll_joint", joint_names[6]);
 
   joint_names = manager.getChainJointNames("not_a_chain");
-  EXPECT_EQ(0, joint_names.size());
+  EXPECT_EQ(static_cast<size_t>(0), joint_names.size());
 
   joint_names = manager.getChainJointNames("head");
-  EXPECT_EQ(2, joint_names.size());
+  EXPECT_EQ(static_cast<size_t>(2), joint_names.size());
   EXPECT_EQ("head_pan_joint", joint_names[0]);
   EXPECT_EQ("head_tilt_joint", joint_names[1]);
